# Make IntObjectHashMap extend Map