--- a/common/src/main/templates/io/netty/util/collection/KObjectHashMap.template
+++ b/common/src/main/templates/io/netty/util/collection/KObjectHashMap.template
@@ -61,9 +61,9 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
 
     private final Set<@O@> keySet = new KeySet();
     private final Set<Entry<@O@, V>> entrySet = new EntrySet();
-    private final Iterable<PEntry<V>> entries = new Iterable<PEntry<V>>() {
+    private final Iterable<PrimitiveEntry<V>> entries = new Iterable<PrimitiveEntry<V>>() {
         @Override
-        public Iterator<PEntry<V>> iterator() {
+        public Iterator<PrimitiveEntry<V>> iterator() {
             return new PrimitiveIterator();
         }
     };
@@ -153,6 +153,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
     public void putAll(Map<? extends @O@, ? extends V> sourceMap) {
         if (sourceMap instanceof @K@ObjectHashMap) {
             // Optimization - iterate through the arrays.
+            @SuppressWarnings("unchecked")
             @K@ObjectHashMap<V> source = (@K@ObjectHashMap<V>) sourceMap;
             for (int i = 0; i < source.values.length; ++i) {
                 V sourceValue = source.values[i];
@@ -205,6 +206,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
 
     @Override
     public boolean containsValue(Object value) {
+        @SuppressWarnings("unchecked")
         V v1 = toInternal((V) value);
         for (V v2 : values) {
             // The map supports null values; this will be matched as NULL_VALUE.equals(NULL_VALUE).
@@ -216,7 +218,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
     }
 
     @Override
-    public Iterable<PEntry<V>> entries() {
+    public Iterable<PrimitiveEntry<V>> entries() {
         return entries;
     }
 
@@ -463,7 +465,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
                 for (;;) {
                     if (values[index] == null) {
                         keys[index] = oldKey;
-                        values[index] = toInternal(oldVal);
+                        values[index] = oldVal;
                         break;
                     }
 
@@ -498,8 +500,9 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
 
     /**
      * Helper method called by {@link #toString()} in order to convert a single map key into a string.
+     * This is protected to allow subclasses to override the appearance of a given key.
      */
-    private String keyToString(@k@ key) {
+    protected String keyToString(@k@ key) {
         return @O@.toString(key);
     }
 
@@ -540,8 +543,8 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
         @Override
         public boolean retainAll(Collection<?> retainedKeys) {
             boolean changed = false;
-            for(Iterator<PEntry<V>> iter = entries().iterator(); iter.hasNext(); ) {
-                PEntry<V> entry = iter.next();
+            for(Iterator<PrimitiveEntry<V>> iter = entries().iterator(); iter.hasNext(); ) {
+                PrimitiveEntry<V> entry = iter.next();
                 if (!retainedKeys.contains(entry.key())) {
                     changed = true;
                     iter.remove();
@@ -581,7 +584,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
     /**
      * Iterator over primitive entries. Entry key/values are overwritten by each call to {@link #next()}.
      */
-    private final class PrimitiveIterator implements Iterator<PEntry<V>>, PEntry<V> {
+    private final class PrimitiveIterator implements Iterator<PrimitiveEntry<V>>, PrimitiveEntry<V> {
         private int prevIndex = -1;
         private int nextIndex = -1;
         private int entryIndex = -1;
@@ -603,7 +606,7 @@ public class @K@ObjectHashMap<V> implements @K@ObjectMap<V> {
         }
 
         @Override
-        public PEntry<V> next() {
+        public PrimitiveEntry<V> next() {
             if (!hasNext()) {
                 throw new NoSuchElementException();
             }