… codec-http2/src/test/java/io/netty/handler/codec/http2/DefaultHttp2RemoteFlowControllerTest.java
… codec-http2/src/test/java/io/netty/handler/codec/http2/Http2SettingsTest.java
… common/src/test/templates/io.netty.util.collection/KObjectHashMapTest.template
… transport-native-epoll/src/main/java/io/netty/channel/epoll/EpollEventLoop.java