@buchgr 2015-07-14
@nmittler why the `p` prefix ... overloading should suffice?

```java
class IntObjectHashMap {
  containsKey(int key);
  containsKey(Integer key);
}
```