@nmittler 2015-07-14
@buchgr Agreed - done.  I believe I had gone the `pXXX` methods because there were several code paths that were triggering autoboxing due to the value parameter, leading to unambiguous methods.

```java
mymap = new IntObjectHashmap<Long>();
int key = ...
long value = ...

// This is ambiguous because value triggers autoboxing.
mymap.put(key, value);
```

This is better handled by using `valueOf` or explicit casting of the `value` parameter.