@buchgr 2015-07-14
> This is better handled by using valueOf or explicit casting of the value parameter.

ur example should trigger the primitive overload, so all good?