@buchgr 2015-07-14
@nmittler looks good. just some minor comments.