@nmittler 2015-07-14
@buchgr finished another pass.