@nmittler 2015-07-15
@normanmaurer @Scottmitch do you want to take a look?