@nmittler 2015-07-17
Just so this doesn't get lost in outdated comments, Scott was concerned about performance with removing the primitive entry:

> Sry I'm a bit late to the discussion on this one. Do we have any benchmarking to determine if getting rid of this is "acceptable"? I agree it makes things simpler/cleaner, but wondering what getting rid of the "native" entry will cost us.

...

> One interesting place where we iterate over the entries is in DefaultHttp2RemoteFlowController every time we allocate bytes. It would be good to understand the impact in this case.