@nmittler 2015-07-17
@Scottmitch I believe I've addressed all of your comments except performance.  I'll look at that next week.