@normanmaurer 2015-07-17
@nmittler thanks buddy!



> Am 17.07.2015 um 23:48 schrieb Nathan Mittler <notifications@github.com>:
> 
> @Scottmitch I believe I've addressed all of your comments except performance. I'll look at that next week.
> 
> —
> Reply to this email directly or view it on GitHub.
>