@nmittler 2015-07-20
@Scottmitch @normanmaurer 
Ok, I ran the same grpc benchmarks as I did for @Scottmitch's `isWritable` change. There seems to be no noticeable difference:

Before the change:
```
Benchmark                                                          (channelCount)  (clientExecutor)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt       Score        Error  Units
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                       1           SMALL  thrpt   10  551757.180 ± 148116.802  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                       2           SMALL  thrpt   10  638856.590 ±   6682.473  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                      10           SMALL  thrpt   10  559230.658 ±   8026.972  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                     100           SMALL  thrpt   10  408613.526 ±  76460.835  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   10  628036.193 ±   6168.349  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       2           SMALL  thrpt   10  619947.375 ±   6950.683  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                      10           SMALL  thrpt   10  590959.888 ±   5729.453  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                     100           SMALL  thrpt   10  522570.703 ±  14940.326  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                       1           SMALL  thrpt   10  903008.711 ±  66474.153  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                       2           SMALL  thrpt   10  850988.549 ± 134336.759  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                      10           SMALL  thrpt   10  867189.803 ± 142575.778  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                     100           SMALL  thrpt   10  664592.911 ± 113271.438  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                       1           SMALL  thrpt   10  942659.337 ± 155346.451  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                       2           SMALL  thrpt   10  976515.074 ±  33647.525  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                      10           SMALL  thrpt   10  853530.134 ± 172174.175  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                     100           SMALL  thrpt   10  711311.819 ±  25608.369  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                       1           SMALL  thrpt   10  976084.789 ±  76686.123  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                       2           SMALL  thrpt   10  967162.543 ±  41804.845  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                      10           SMALL  thrpt   10  982240.506 ±  53122.231  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                     100           SMALL  thrpt   10       ≈ 10⁰               ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                       1           SMALL  thrpt   10  900843.634 ±  57505.936  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                       2           SMALL  thrpt   10  990676.809 ±  58005.589  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                      10           SMALL  thrpt   10  922337.886 ±  57449.446  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                     100           SMALL  thrpt   10       ≈ 10⁰               ops/s
```

After the change:
```
Benchmark                                                          (channelCount)  (clientExecutor)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt       Score        Error  Units
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                       1           SMALL  thrpt   10  642027.088 ±  35936.969  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                       2           SMALL  thrpt   10  596118.362 ±  25066.277  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                      10           SMALL  thrpt   10  583812.430 ±  22636.985  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1           DEFAULT                     100           SMALL  thrpt   10  495768.094 ±   9691.953  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   10  641037.809 ±   6559.519  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       2           SMALL  thrpt   10  600069.933 ±  68953.616  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                      10           SMALL  thrpt   10  613251.185 ±   5141.680  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                     100           SMALL  thrpt   10  521176.137 ±  51090.564  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                       1           SMALL  thrpt   10  753689.952 ± 122044.652  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                       2           SMALL  thrpt   10  875496.862 ±  94797.461  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                      10           SMALL  thrpt   10  808643.298 ± 124313.169  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2           DEFAULT                     100           SMALL  thrpt   10  643546.856 ±  74414.465  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                       1           SMALL  thrpt   10  802412.420 ± 138024.696  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                       2           SMALL  thrpt   10  972859.341 ±  64605.156  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                      10           SMALL  thrpt   10  991742.924 ±  45957.893  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               2            DIRECT                     100           SMALL  thrpt   10  658660.355 ±  71239.538  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                       1           SMALL  thrpt   10  941348.671 ±  39534.913  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                       2           SMALL  thrpt   10  977336.631 ±  43517.410  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                      10           SMALL  thrpt   10  927061.599 ±  28209.328  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4           DEFAULT                     100           SMALL  thrpt   10       ≈ 10⁰               ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                       1           SMALL  thrpt   10  890787.413 ±  37046.598  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                       2           SMALL  thrpt   10  937797.715 ±  44238.085  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                      10           SMALL  thrpt   10  965742.144 ±  58288.124  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               4            DIRECT                     100           SMALL  thrpt   10       ≈ 10⁰               ops/s
```